rev V2

+++ resources/views/roles/index.blade.php
@@ -1,47 +1,56 @@
-@extends('layouts.app-master')
+@extends('layouts.app')
+
 
 @section('content')
-    
-    
-    <div class="bg-light p-4 rounded">
-        <h1>Roles</h1>
-        <div class="lead">
-            Manage your roles here.
-            <a href="{{ route('roles.create') }}" class="btn btn-primary btn-sm float-right">Add role</a>
+<div class="row">
+    <div class="col-lg-12 margin-tb">
+        <div class="pull-left">
+            <h2>Role Management</h2>
         </div>
-        
-        <div class="mt-2">
-            @include('layouts.partials.messages')
+        <div class="pull-right">
+        @can('role-create')
+            <a class="btn btn-success" href="{{ route('roles.create') }}"> Create New Role</a>
+            @endcan
         </div>
-
-        <table class="table table-bordered">
-          <tr>
-             <th width="1%">No</th>
-             <th>Name</th>
-             <th width="3%" colspan="3">Action</th>
-          </tr>
-            @foreach ($roles as $key => $role)
-            <tr>
-                <td>{{ $role->id }}</td>
-                <td>{{ $role->name }}</td>
-                <td>
-                    <a class="btn btn-info btn-sm" href="{{ route('roles.show', $role->id) }}">Show</a>
-                </td>
-                <td>
-                    <a class="btn btn-primary btn-sm" href="{{ route('roles.edit', $role->id) }}">Edit</a>
-                </td>
-                <td>
-                    {!! Form::open(['method' => 'DELETE','route' => ['roles.destroy', $role->id],'style'=>'display:inline']) !!}
-                    {!! Form::submit('Delete', ['class' => 'btn btn-danger btn-sm']) !!}
-                    {!! Form::close() !!}
-                </td>
-            </tr>
-            @endforeach
-        </table>
-
-        <div class="d-flex">
-            {!! $roles->links() !!}
-        </div>
-
     </div>
+</div>
+
+
+@if ($message = Session::get('success'))
+    <div class="alert alert-success">
+        <p>{{ $message }}</p>
+    </div>
+@endif
+
+
+<table class="table table-bordered">
+  <tr>
+     <th>No</th>
+     <th>Name</th>
+     <th width="280px">Action</th>
+  </tr>
+    @foreach ($roles as $key => $role)
+    <tr>
+        <td>{{ ++$i }}</td>
+        <td>{{ $role->name }}</td>
+        <td>
+            <a class="btn btn-info" href="{{ route('roles.show',$role->id) }}">Show</a>
+            @can('role-edit')
+                <a class="btn btn-primary" href="{{ route('roles.edit',$role->id) }}">Edit</a>
+            @endcan
+            @can('role-delete')
+                {!! Form::open(['method' => 'DELETE','route' => ['roles.destroy', $role->id],'style'=>'display:inline']) !!}
+                    {!! Form::submit('Delete', ['class' => 'btn btn-danger']) !!}
+                {!! Form::close() !!}
+            @endcan
+        </td>
+    </tr>
+    @endforeach
+</table>
+
+
+{!! $roles->render() !!}
+
+
+<p class="text-center text-primary"><small>Tutorial by ItSolutionStuff.com</small></p>
 @endsection
